% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDESeq.R
\docType{package}
\name{robustDESeq-package}
\alias{robustDESeq}
\alias{robustDESeq-package}
\title{robustDESeq}
\description{
\code{robustDESeq} is a package for robust negative binomial regression for genomics data analysis
}
\author{
\strong{Maintainer}: Tim Barry \email{tbarry@hsph.harvard.edu}

}
\keyword{package}
