% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_deseq2.R
\name{run_deseq_list_interface}
\alias{run_deseq_list_interface}
\title{Run robust DESeq (list interface)}
\usage{
run_deseq_list_interface(
  Y_list,
  x,
  Z,
  side = "two_tailed",
  h = 15L,
  alpha = 0.1,
  robust = TRUE,
  max_iterations = 200000L
)
}
\arguments{
\item{Y_list}{a list of response vectors}

\item{x}{the treatment vector}

\item{Z}{the covariate matrix}

\item{side}{the side of the test, one of "left", "right", or "two_tailed"}

\item{h}{tuning parameter for the anytime-valid permutation test}

\item{alpha}{nominal FDR of the BH procedure}
}
\value{
results data frame
}
\description{
Run robust DESeq (list interface)
}
\examples{
n <- 100L
m <- 500L
theta_gt <- 5
Z <- MASS::mvrnorm(n = n, mu = c(-0.5, 0.5), Sigma = toeplitz(c(1, 0.5)))
x <- rbinom(n = n, size = 1, prob = 0.3)
family_object <- MASS::negative.binomial(theta_gt)
design_matrix <- cbind(x, Z)
colnames(design_matrix) <- c("x", "z1", "z2")
null_coefs <- log(c(25, 1.0, 0.8, 1.1))
alt_coefs <- log(c(25, 1.5, 0.8, 1.1))
under_null <- sample(c(rep(TRUE, 0.9 * m), rep(FALSE, 0.1 * m)))
Y_list <- sapply(X = seq_len(m), FUN = function(i) {
 generate_glm_data(design_matrix = design_matrix,
                   coefficients = if (under_null[i]) null_coefs else alt_coefs,
                   family_object = family_object,
                   add_intercept = TRUE)
}, simplify = FALSE)
h <- 15L; alpha <- 0.1; theta <- 30

# mass implementation
res_robust_nb <- run_robust_nb_regression(Y_list, x, Z)
get_result_metrics(res_robust_nb, under_null)

# deseq implementation
res_deseq <- run_deseq_list_interface(Y_list, x, Z, robust = FALSE)
get_result_metrics(res_deseq, under_null)
}
