% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_regression_methods.R
\name{run_linear_model}
\alias{run_linear_model}
\title{Run linear model}
\usage{
run_linear_model(Y_list, x, Z, side = "two_tailed", alpha = 0.1)
}
\arguments{
\item{Y_list}{a list of response vectors}

\item{x}{the treatment vector}

\item{Z}{the covariate matrix}

\item{side}{the side of the test, one of "left", "right", or "two_tailed"}

\item{alpha}{nominal FDR of the BH procedure}
}
\value{
a data frame containing the p-values and rejections
}
\description{
Run linear model
}
