% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_methods.R
\name{run_mann_whitney_test_simple}
\alias{run_mann_whitney_test_simple}
\title{Run Mann-Whitney test}
\usage{
run_mann_whitney_test_simple(s1, s2)
}
\arguments{
\item{x}{binary vector of treatment indicators}

\item{y}{observations for the second sample}
}
\value{
the z-score
}
\description{
Run Mann-Whitney test
}
\examples{
# x <- rbinom(n = 500, size = 1, prob = 0.5)
# y <- MASS::rnegbin(n = n, mu = 50, theta = 10)
n <- 1000
# s <- rnorm(n = n, mean = 50)
s <- MASS::rnegbin(n = n, mu = 40, theta = 10)
x <- s[1:500]
y <- s[501:1000]
run_mann_whitney_test_simple(x, y)
wilcox.test(x, y)
}
