% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_deseq2.R
\name{run_robust_deseq}
\alias{run_robust_deseq}
\title{Title}
\usage{
run_robust_deseq(
  dds,
  side = "two_tailed",
  h = 15L,
  alpha = 0.1,
  max_iterations = 200000L
)
}
\arguments{
\item{max_iterations}{}
}
\description{
Title
}
\examples{
library(DESeq2)
dds <- readRDS("/Users/tib163/research_offsite/projects/camp/data/arrayed_crispr/deseq_object.rds")
design(dds) <- formula(~donor_id + stim_status + grna)
row_sums <- rowSums(assays(dds)$counts)
highly_expressed <- row_sums >= 6
dds <- dds[highly_expressed,]
res <- run_robust_deseq(dds)
}
