% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_robust_vgam.R
\name{run_robust_vgam}
\alias{run_robust_vgam}
\title{Run robust NB regression}
\usage{
run_robust_vgam(
  y_list,
  x,
  Z_list,
  side = "two_tailed",
  h = 15L,
  alpha = 0.1,
  max_iterations = 200000L,
  custom_permutation_list = list()
)
}
\arguments{
\item{x}{the treatment vector}

\item{side}{the side of the test, one of "left", "right", or "two_tailed"}

\item{h}{tuning parameter for the anytime-valid permutation test}

\item{alpha}{nominal FDR of the BH procedure}

\item{Y_list}{a list of response vectors}

\item{Z}{the covariate matrix}

\item{method}{the method to use to carry out the NB regression under the null model, either "MASS" or "VGAM"}
}
\value{
a data frame containing the rejection set
}
\description{
Run robust NB regression
}
\examples{
######################
# GENERATE ONE DATASET
######################
library(VGAM)
set.seed(10)
n <- 5000L
m <- 100L
theta_gt <- 5
x <- rbinom(n = n, size = 1, prob = 0.3)
null_coefs <- log(c(7, 1.0, 0.8, 1.1))
alt_coefs <- log(c(7, 1.1, 0.8, 1.1))
under_null <- sample(c(rep(TRUE, 0.9 * m), rep(FALSE, 0.1 * m)))
dat_list <- lapply(seq_len(100), function(i) {
  Z <- MASS::mvrnorm(n = n, mu = c(-0.5, 0.5), Sigma = toeplitz(c(1, 0.5)))
  colnames(Z) <- c("1", "2")
  family_object <- MASS::negative.binomial(theta_gt)
  design_matrix <- cbind(x, Z)
  colnames(design_matrix) <- c("x", "z1", "z2")
  y <- generate_glm_data(design_matrix = design_matrix,
      coefficients = if (under_null[i]) null_coefs else alt_coefs,
      family_object = family_object,
      add_intercept = TRUE
    )
    list(Z = Z, y = y)
})
y_list <- lapply(X = dat_list, FUN = function(l) l[["y"]])
Z_list <- lapply(X = dat_list, FUN = function(l) l[["Z"]])

# standard VGAM
robust_res <- run_robust_vgam(y_list, x, Z_list)
get_result_metrics(robust_res, under_null)
}
