% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_methods.R
\name{run_standard_nb_regression}
\alias{run_standard_nb_regression}
\title{Run standard NB regression}
\usage{
run_standard_nb_regression(
  Y_list,
  x,
  Z,
  side = "two_tailed",
  alpha = 0.1,
  method = "MASS"
)
}
\arguments{
\item{Y_list}{a list of response vectors}

\item{x}{the treatment vector}

\item{Z}{the covariate matrix}
}
\value{
a data frame containing the p-values and rejections
}
\description{
Run standard NB regression
}
