% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_test.R
\name{run_mann_whitney_test_permutations}
\alias{run_mann_whitney_test_permutations}
\title{Run Mann-Whitney test (finite-sample)}
\usage{
run_mann_whitney_test_permutations(
  Y_list,
  x,
  Z,
  side = "two_tailed",
  alpha = 0.1
)
}
\arguments{
\item{Y_list}{a list of response vectors}

\item{x}{the treatment vector}

\item{Z}{the covariate matrix}

\item{side}{the side of the test, one of "left", "right", or "two_tailed"}

\item{alpha}{nominal FDR of the BH procedure}
}
\value{
a data frame containing the results
}
\description{
Run Mann-Whitney test (finite-sample)
}
