% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_methods.R
\name{run_regress_out_covariates_test}
\alias{run_regress_out_covariates_test}
\title{Title}
\usage{
run_regress_out_covariates_test(
  Y_list,
  x,
  Z,
  h = 15L,
  alpha = 0.1,
  resid_type = c("response", "deviance", "pearson")[1]
)
}
\arguments{
\item{resid_type}{}
}
\description{
Title
}
