% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_methods.R
\name{run_regress_out_covariates_test}
\alias{run_regress_out_covariates_test}
\title{Run regress out residuals test}
\usage{
run_regress_out_covariates_test(
  Y_list,
  x,
  Z,
  side = "two_tailed",
  h = 15L,
  alpha = 0.1,
  resid_type = c("response", "deviance", "pearson")[1]
)
}
\arguments{
\item{Y_list}{a list of response vectors}

\item{x}{the treatment vector}

\item{Z}{the covariate matrix}

\item{side}{the side of the test, one of "left", "right", or "two_tailed"}

\item{h}{tuning parameter for the anytime-valid permutation test}

\item{alpha}{nominal FDR of the BH procedure}
}
\value{
a data frame containing the p-values and rejections
}
\description{
Run regress out residuals test
}
