% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_deseq2.R
\name{run_robust_deseq}
\alias{run_robust_deseq}
\title{Run robust DESeq}
\usage{
run_robust_deseq(
  dds,
  side = "two_tailed",
  h = 15L,
  alpha = 0.1,
  size_factors = NULL,
  size_factor_estimation = "default",
  dispersions = NULL,
  dispersion_estimation = "local",
  max_iterations = 200000L,
  custom_permutation_list = list(),
  precomp_list = NULL,
  return_precomp = FALSE
)
}
\arguments{
\item{dds}{a DESeq object}

\item{side}{the side of the test, one of "left", "right", or "two_tailed"}

\item{h}{tuning parameter for the anytime-valid permutation test}

\item{alpha}{nominal FDR of the BH procedure}
}
\value{
the result data frame
}
\description{
Run robust DESeq
}
\examples{
library(DESeq2)
dds <- readRDS("/Users/tib163/research_offsite/projects/camp/data/arrayed_crispr/deseq_object.rds")
design(dds) <- formula(~donor_id + stim_status + grna)
row_sums <- rowSums(assays(dds)$counts)
highly_expressed <- row_sums >= 5
dds <- dds[highly_expressed,]
res <- run_robust_deseq(dds)
}
