% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_robust_nb_glm.R
\name{run_robust_nb_regression}
\alias{run_robust_nb_regression}
\title{Run robust NB regression}
\usage{
run_robust_nb_regression(Y_list, x, Z, h = 15L, alpha = 0.1, theta = NULL)
}
\arguments{
\item{Y_list}{a list of response vectors}

\item{x}{the treatment vector}

\item{Z}{the covariate matrix}
}
\value{
a data frame containing the rejection set
}
\description{
Run robust NB regression
}
\examples{
############################
# GENERATE MULTIPLE DATASETS
############################
n <- 1000L
m <- 500L
x <- rbinom(n = n, size = 1, prob = 0.25)
Z <- MASS::mvrnorm(n = n, mu = c(-0.5, 0.5), Sigma = toeplitz(c(1, 0.5)))
colnames(Z) <- c("1", "2")
family_object <- MASS::negative.binomial(5)
design_matrix <- cbind(x, Z)
colnames(design_matrix) <- c("x", "z1", "z2")
null_coefs <- log(c(7, 1.0, 0.8, 1.1))
alt_coefs <- log(c(7, 1.3, 0.8, 1.1))
under_null <- sample(size = m, c(TRUE, FALSE), replace = TRUE, prob = c(0.95, 0.05))
Y_list <- sapply(X = seq_len(m), FUN = function(i) {
  generate_glm_data(design_matrix = design_matrix,
  coefficients = if (under_null[i]) null_coefs else alt_coefs,
  family_object = family_object,
  add_intercept = TRUE)
}, simplify = FALSE)
h <- 15L; alpha <- 0.1; theta <- NULL

########################
# STANDARD NB REGRESSION
########################
standard_res <- run_standard_nb_regression(Y_list = Y_list, x = x, Z = Z)
mean(under_null[standard_res |> dplyr::filter(rejected) |> dplyr::pull(hyp_idx)])

######################
# ROBUST NB REGRESSION
######################
robust_res <- run_robust_nb_regression(Y_list = Y_list, x = x, Z = Z)
mean(under_null[robust_res |> dplyr::filter(rejected) |> dplyr::pull(hyp_idx)])
}
