% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_deseq2.R
\name{run_standard_deseq_list_interface}
\alias{run_standard_deseq_list_interface}
\title{Run standard DESeq (list interface)}
\usage{
run_standard_deseq_list_interface(
  Y_list,
  x,
  Z,
  side = "two_tailed",
  alpha = 0.1,
  dispersion_estimation = "local",
  size_factors = NULL
)
}
\description{
Run standard DESeq (list interface)
}
