% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_regression_methods.R
\name{run_standard_nb_regression}
\alias{run_standard_nb_regression}
\title{Run standard NB regression}
\usage{
run_standard_nb_regression(
  Y_list,
  x,
  Z,
  side = "two_tailed",
  alpha = 0.1,
  method = "MASS",
  theta = NULL
)
}
\arguments{
\item{Y_list}{a list of response vectors}

\item{x}{the treatment vector}

\item{Z}{the covariate matrix}

\item{side}{the side of the test, one of "left", "right", or "two_tailed"}

\item{alpha}{nominal FDR of the BH procedure}

\item{method}{the method to use to carry out the NB regression under the null model, either "MASS" or "VGAM"}
}
\value{
a data frame containing the p-values and rejections
}
\description{
Run standard NB regression
}
